Fix windows build with SCons amd remove more warnings

+++ contrib/partio4Maya/partioEmitter.cpp
@@ -183,7 +183,7 @@ MStatus partioEmitter::initialize()
         eAttr.addField(formatExtMap[i].toUpperCase(),	i);
     }
 
-    eAttr.setDefault(Partio::readFormatIndex("pdc")); // PDC
+    eAttr.setDefault(short(Partio::readFormatIndex("pdc"))); // PDC
     eAttr.setChannelBox(true);
     eAttr.setKeyable(false);
